BE/requirements.txt
botocore==1.12.97
Click==7.0
docutils==0.14
ecdsa==0.13
Flask==1.0.2
Flask-Cors==3.0.6
future==0.17.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonpickle==1.1
MarkupSafe==1.1.0
psycopg2-binary==2.7.7
pyasn1==0.4.5
python-dateutil==2.8.0
python-jose==3.0.1
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
